--- a/maps/biter_battles_v2/terrain.lua
+++ b/maps/biter_battles_v2/terrain.lua
@@ -8,6 +8,7 @@ local AiTargets = require "maps.biter_battles_v2.ai_targets"
 local tables = require "maps.biter_battles_v2.tables"
 local session = require 'utils.datastore.session_data'
 local biter_texture = require "maps.biter_battles_v2.biter_texture"
+local river = require "maps.biter_battles_v2.precomputed.river"
 
 local spawn_ore = tables.spawn_ore
 local table_insert = table.insert
@@ -156,17 +157,18 @@ local function is_within_spawn_island(pos)
 	return true
 end
 
--- border_river_noise is the maximum random value that can be added to each side of the river
-local border_river_noise = 4
-local river_width_half_min = math_floor(bb_config.border_river_width * -0.5)
-local river_width_half_max = river_width_half_min - border_river_noise
--- pos must be from the North side
-local function is_horizontal_border_river(pos)
-	if tile_distance_to_center(pos) < river_circle_size then return true end
-	if pos.y < river_width_half_max then return false end
-	if pos.y > river_width_half_min then return true end
-	if pos.y >= river_width_half_min - (math_abs(Functions.get_noise(1, pos)) * border_river_noise) then return true end
-	return false
+local river_width_half = math_floor(bb_config.border_river_width * 0.5)
+local function is_horizontal_border_river(pos, seed)
+	if tile_distance_to_center(pos) < river_circle_size then
+		return true
+	end
+
+	-- Offset contains coefficient in a range between [-4,4]. Select it
+	-- from pre-computed table. Position X coordinate is used to determinate
+	-- which offset is selected.
+	local offset = river.offset[(pos.x + seed) % river.size]
+	local y = -(pos.y + offset)
+	return (y <= river_width_half)
 end
 
 local function generate_starting_area(pos, surface)
@@ -174,7 +176,8 @@ local function generate_starting_area(pos, surface)
 	local noise_multiplier = 15
 	local min_noise = -noise_multiplier * 1.25
 
-	if is_horizontal_border_river(pos) then
+	local seed = game.surfaces[global.bb_surface_name].map_gen_settings.seed
+	if is_horizontal_border_river(pos, seed) then
 		return
 	end
 
@@ -257,10 +260,11 @@ end
 
 local function generate_river(surface, left_top_x, left_top_y)
 	if not (left_top_y == -32 or (left_top_y == -64 and (left_top_x == -32 or left_top_x == 0))) then return end
+	local seed = game.surfaces[global.bb_surface_name].map_gen_settings.seed
 	for x = 0, 31, 1 do
 		for y = 0, 31, 1 do
 			local pos = {x = left_top_x + x, y = left_top_y + y}
-			if is_horizontal_border_river(pos) and not is_within_spawn_island(pos) then
+			if is_horizontal_border_river(pos, seed) and not is_within_spawn_island(pos) then
 				surface.set_tiles({{name = "deepwater", position = pos}})
 				if global.random_generator(1, 64) == 1 then
 					local e = surface.create_entity({name = "fish", position = pos})
@@ -692,11 +696,12 @@ end
 
 --Landfill Restriction
 function Public.restrict_landfill(surface, user, tiles)
+	local seed = game.surfaces[global.bb_surface_name].map_gen_settings.seed
 	for _, t in pairs(tiles) do
 		local check_position = t.position
 		if check_position.y > 0 then check_position = {x = check_position.x, y = (check_position.y * -1) - 1} end
 		local trusted = session.get_trusted_table()
-		if is_horizontal_border_river(check_position) then
+		if is_horizontal_border_river(check_position, seed) then
 			surface.set_tiles({{name = t.old_tile.name, position = t.position}}, true)
 			if user ~= nil then
 				user.print('You can not landfill the river', {r = 0.22, g = 0.99, b = 0.99})