… maps/biter_battles_v2/precomputed/biter_texture.lua
… maps/biter_battles_v2/precomputed/chunk_container.lua
… maps/biter_battles_v2/precomputed/river.lua