### Brief description of the changes:
Optimized river generation by reducing amount of API calls, better memory management and lookup table with pseudo random offset values.

#### Current performance:
190x "generate_river" Total Duration: 2949.216225ms
Average time spent in function: 15.5ms

#### With new changes applied (just static buffer)
134x "generate_river" Total Duration: 1359.028429ms
Average time spent in function: 10.1ms

#### With new changes applied (full memory rework)
237x "generate_river" Total Duration: 1754.767363ms
Average time spent in function: 7ms

### Tested Changes:
- [X] I've tested the changes locally.
- [ ]  I've tested the changes online.