@XVhc6A 2023-10-25
Could you provide a brief description + script that produces these precomputed files?  I'm okay with having them, but if someone wants to change it later it shouldn't be so black-box.