@developer-8 2023-11-01
![Screenshot from 2023-11-02 00-12-48](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/88335092/b749db54-4d2d-466e-86b3-be0e227325ba)
1. After this PR, every map has the same coastline pattern
2. The shape of the coastline is different. The original is smoother.