@developer-8 2023-11-02
3.
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/88335092/fb9524a7-603b-4dfc-a51b-3e2430d82487)
Original river variate water/land on 4 tiles. y for north side: -26, -25, -24, -23
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/88335092/8a25ea7f-0314-41b9-bc76-ec96db8dc465)
After this PR it variate on 2 tiles. y for north side: -23, -22
This also means that the river has shrunk by 2 tiles.

4.`generate_river` and as consequence `is_horizontal_border_river` doing work only 1 time for 1 chunk near river so I think its acceptable sacrifice 5.4ms (from 15.5ms to 10.1ms) performance gain and have unique river every game