@cogito123 2023-11-05
I'll need to work some more on the pattern to make it smoother. But it's now seeded.