@cogito123 2023-11-06
> Could you provide a brief description + script that produces these precomputed files? I'm okay with having them, but if someone wants to change it later it shouldn't be so black-box.

I totally agree, current feedback from discord is that I should create separate repo for tool that I use for these precomputed files. At the moment there are several executables for that. I started work on merging all of these into one. Once I have it done, I'll publish it under my name. In the meantime if you propose to do it differently or host it somewhere in this repo, I'm also down for that. Keep in mind, this is program written in rust language/