@cogito123 2023-11-06
I also reduced size of chunk_container.lua and river.lua, by dropping unnecessary index as suggested by @developer-8 and moved biter_texture to precomputed directory to make it more organized