@developer-8 2023-11-09
Original river variate water/land on 4 tiles. y for north side: -26, -25, -24, -23

After the PR is in the current state it variate on 4 tiles. y for north side: -28, -27, -26, -25
This also means that the river has expanded by 4 tiles.
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/88335092/7f410bdd-86b0-44e1-bc72-782c4b790929)