@cogito123 2023-11-19
Thanks for double checking. I remember checking it myself, but perhaps I changed something to pattern itself after, which made it offset unexpectedly.