@developer-8 2023-11-19
`[ERROR] ...temp/currently-playing/maps/biter_battles_v2/terrain.lua:170: attempt to perform arithmetic on local 'seed' (a nil value) in .../temp/currently-playing/maps/biter_battles_v2/main.lua:115`
probably caused by no seed passed to function
 https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/blob/76d52922d3f98f168914ccba1fc8c275b695c898/maps/biter_battles_v2/terrain.lua#L699