@Ragnarok77-factorio 2023-11-19
Thanks for quickfix and info @cogito123 and @developer-8 . server updated for main, should be fine there, for captain it will be another time